+++ fusion/src/Fusion/python/helpers.h
@@ -1,7 +1,7 @@
 // helpers.h
 #pragma once
 
-#include "../tensor/tensor.h"
+#include "../tensor.h"
 #include <numeric>
 #include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
